Clean up wayland device jail
The jail for the wayland device used chown to ensure that its jail had
the proper permissions for the wayland socket to be bind mounted into
it.  This creates some unnecessary complexity because it requires
careful management of the user and group and crosvm runs as (a non-root
user cannot change the owner of a directory) or that crosvm has the
CAP_CHOWN capability.

Instead of trying to make the permissions fit, just have the jail mount
a small tmpfs over the jail's root directory.  This is one of the things
that a process inside a user namespace has the ability to do.  Bind
mounting the wayland socket into this tmpfs then just works without any
other issues.

BUG=chromium:799523
TEST=linux vm boots with no errors

Change-Id: Ic2240f430c7fd332a15b4fcd4e52374799eb6c9d
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/851413
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>